backend/requirements.txt
fastapi
uvicorn
pydantic
google-cloud-storage
google-cloud-bigquery
joblib